# backend/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Worship Core Backend
  version: 0.1.0
  description: Server-side session authentication; mutations require the CSRF header obtained from /api/auth/csrf.
paths:
...
  /api/workspaces/{workspaceId}/documents/{documentId}/setlist:
    get:
      operationId: getSetlist
      description: Current authorized Setlist snapshot and aggregate version
      security:
      - session: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetlistServiceSetlistViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    put:
      operationId: updateSetlist
      description: Title and notes; expectedVersion required
      security:
      - session: []
        csrf: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetlistServiceSetlistViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetlistControllerUpdateRequest'
  /api/workspaces/{workspaceId}/documents/{documentId}/setlist/items:
    post:
      operationId: addSetlistItem
      description: Select workspace songId; expectedVersion required
      security:
      - session: []
        csrf: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetlistServiceSetlistViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetlistControllerSongRequest'
  /api/workspaces/{workspaceId}/documents/{documentId}/setlist/items/{itemId}:
    delete:
      operationId: removeSetlistItem
      description: Remove and renumber atomically; expectedVersion query parameter required
      security:
      - session: []
        csrf: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetlistServiceSetlistViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: itemId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: expectedVersion
        in: query
        required: true
        schema:
          type: integer
          format: int64
          minimum: 0
  /api/workspaces/{workspaceId}/documents/{documentId}/setlist/items/order:
    put:
      operationId: reorderSetlistItems
      description: itemIds must contain every item once; expectedVersion required
      security:
      - session: []
        csrf: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetlistServiceSetlistViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetlistControllerOrderRequest'
  /api/workspaces/{workspaceId}/documents/{documentId}/setlist/items/{itemId}/song:
    put:
      operationId: selectSetlistSong
      description: Select existing workspace songId; expectedVersion required
      security:
      - session: []
        csrf: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetlistServiceSetlistViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: itemId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetlistControllerSongRequest'
  /api/workspaces/{workspaceId}/documents/{documentId}/setlist/items/{itemId}/settings:
    put:
      operationId: updateMusicalSettings
      description: key, bpm and sessions; expectedVersion required
      security:
      - session: []
        csrf: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetlistServiceSetlistViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: itemId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetlistControllerSettingsRequest'
  /api/workspaces/{workspaceId}/documents/{documentId}/setlist/items/{itemId}/notes:
    put:
      operationId: updateItemNotes
      description: notes and expectedVersion required
      security:
      - session: []
        csrf: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetlistServiceSetlistViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: itemId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetlistControllerNotesRequest'
  /api/workspaces/{workspaceId}/documents/{documentId}/setlist/items/{itemId}/song-form:
    put:
      operationId: updateSongForm
      description: Structured blocks with stable id, section, repeat, cue, calling, note; expectedVersion required
      security:
      - session: []
        csrf: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetlistServiceSetlistViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: itemId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetlistControllerFormRequest'
...
  /api/workspaces/{workspaceId}/documents/{documentId}/setlist/items/{itemId}/score:
    put:
      operationId: selectItemScore
      description: scoreId or null and expectedVersion; tenant and EDIT authorization checked
      security:
      - session: []
        csrf: []
      responses:
        '200':
          description: Capability completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetlistServiceSetlistViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid input or file
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current permission or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Stale source version
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Provider unavailable or not configured
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: itemId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetlistControllerScoreSelectionRequest'
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: SESSION
    csrf:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: Use actual headerName/token from /api/auth/csrf, not a long-lived cached token
  schemas:
...
    SetlistServiceSetlistViewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        documentId:
          type: integer
          format: int64
        title:
          type: string
        version:
          type: integer
          format: int64
        notes:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/SetlistServiceItemViewResponse'
      required:
      - id
      - documentId
      - title
      - version
      - notes
      - items
...
  responses:
    Error400:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - VALIDATION_FAILED
    Error401:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - AUTHENTICATION_REQUIRED
    Error403:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - ACCESS_DENIED
                  - REAUTHENTICATION_REQUIRED
                  - YOUTUBE_AUTHORIZATION_REQUIRED
    Error404:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - NOT_FOUND
                  - REQUEST_REJECTED
    Error405:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - REQUEST_REJECTED
    Error406:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - REQUEST_REJECTED
...